move groups reducer
